Show the methods for calculating a noise threshold

This function outputs the methods available for the calculation of the noise threshold. To be used as input in calculate_noise_threshold_*().

get_methods_calculate_noise_threshold()

Value

A character vector of options for the method arguement of calculate_noise_threshold_*()

Examples

get_methods_calculate_noise_threshold()

#>  [1] "Density_based-No_normalisation"      
#>  [2] "Density_based-RPM_normalisation"     
#>  [3] "Density_based-Quantile_normalisation"
#>  [4] "Line_plot-No_smoothing"              
#>  [5] "Line_plot-loess10_smoothing"         
#>  [6] "Line_plot-loess25_smoothing"         
#>  [7] "Line_plot-loess50_smoothing"         
#>  [8] "Boxplot-Median"                      
#>  [9] "Boxplot-IQR"                         
#> [10] "Boxplot-Quant5"