Show the methods for calculating correlation or distance

This function outputs the methods available for the calculation of the correlation or distance. The standard correlation methods use stats::cor and a wide variety of distance methods are available using the philentropy package. To be used as input in calculate_expression_similarity_*().

get_methods_correlation_distance(names = TRUE)

Arguments

names
whether to output names (default) or characterisation as similarity or dissimilarity (used internally to invert dissimilarity measures)

Value

A character vector of options for the method arguement of calculate_expression_similarity_*(); if names=FALSE, a vector of types (similarity/dissimilarity measure) of the same length

Examples

get_methods_correlation_distance()

#>  [1] "correlation_pearson"        "correlation_kendall"       
#>  [3] "correlation_spearman"       "distance_euclidean"        
#>  [5] "distance_manhattan"         "distance_minkowski"        
#>  [7] "distance_chebyshev"         "distance_sorensen"         
#>  [9] "distance_gower"             "distance_soergel"          
#> [11] "distance_kulczynski_d"      "distance_canberra"         
#> [13] "distance_lorentzian"        "distance_intersection"     
#> [15] "distance_non-intersection"  "distance_wavehedges"       
#> [17] "distance_czekanowski"       "distance_motyka"           
#> [19] "distance_kulczynski_s"      "distance_tanimoto"         
#> [21] "distance_ruzicka"           "distance_inner_product"    
#> [23] "distance_harmonic_mean"     "distance_cosine"           
#> [25] "distance_hassebrook"        "distance_jaccard"          
#> [27] "distance_dice"              "distance_fidelity"         
#> [29] "distance_bhattacharyya"     "distance_hellinger"        
#> [31] "distance_matusita"          "distance_squared_chord"    
#> [33] "distance_squared_euclidean" "distance_pearson"          
#> [35] "distance_neyman"            "distance_squared_chi"      
#> [37] "distance_prob_symm"         "distance_divergence"       
#> [39] "distance_clark"             "distance_additive_symm"    
#> [41] "distance_kullback-leibler"  "distance_jeffreys"         
#> [43] "distance_k_divergence"      "distance_topsoe"           
#> [45] "distance_jensen-shannon"    "distance_jensen_difference"
#> [47] "distance_taneja"            "distance_kumar-johnson"    
#> [49] "distance_avg"