Function to remove the noisy reads from the expression matrix using a specified method

This function is used to remove the noisy reads from the expression matrix. It can use thresholds precalculated by calculate_threshold_noise(), if available. If a vector of thresholds needs to be input manually, use remove_noise_matrix() instead.

remove_noise_method(
  expression.matrix,
  method.chosen = "Boxplot-IQR",
  stats.df = NULL,
  abn.matrix = NULL,
  dist.matrix = NULL,
  dist.thresh = 0.25,
  binsize = 0.1,
  add.thresh = TRUE,
  average.thresh = TRUE,
  remove.noisy.features = TRUE,
  export.csv = NULL
)

Arguments

expression.matrix
the expression matrix

method.chosen
the method to be used for the noise removal; must be one of get_methods_calculate_noise_threshold()

stats.df
a tibble, as output by calculate_threshold_noise(); if supplied and the chosen method appears in it, the corresponding threshold is used directly.

abn.matrix, dist.matrix
the input distance and abundance matrices as calculated by calculate_distance_matrices(); only needed if stats.df is not supplied; if either is not supplied, only a fixed threshold is calculated based on the density

dist.thresh
a distance threshold to be used if the noise thresholds are not pre-calculated; the default 0.25 is suitable for correlation measures

binsize
size of each bin in the boxplot methods; defaults to 0.1 (on a log-scale)

add.thresh
whether to add the noise threshold to all values in the expression matrix (default), or set entries below the threshold to the threshold

average.thresh
if TRUE (default), uses tthe average of the vector of thresholds across all samples; if FALSE, uses the thresholds as supplied

remove.noisy.features
logical, whether rows of the expression matrix that are fully under the noise threshold should be removed (default TRUE)

export.csv
option to write the matrix into a csv after the noise removal; should be NULL or the name of the output file.

Value

Returns a matrix of the same dims as the expression matrix, with the noise removed using the specified method. All entries below the noise threshold are replaced with the noise threshold.

Examples

obj <- calculate_distance_matrices_counts(
    expression.matrix = matrix(1:100, ncol=5),
    method="correlation_pearson",
    n.elements.per.window=3)

#> The input matrix has 20 rows and 5 cols

#>     number of genes: 20

#>     number of samples: 5

#>     the number of elements per window is 3

#>     the step size is 1

#> Method chosen: correlation_pearson

#> Working with sample 1

#> Working with sample 2

#> Working with sample 3

#> Working with sample 4

#> Working with sample 5

remove_noise_method(
    expression.matrix=obj$exp,
    abn.matrix=obj$abn,
    dist.matrix=obj$dist)

#>       [,1] [,2] [,3] [,4] [,5]
#>  [1,]    2   22   42   62   82
#>  [2,]    3   23   43   63   83
#>  [3,]    4   24   44   64   84
#>  [4,]    5   25   45   65   85
#>  [5,]    6   26   46   66   86
#>  [6,]    7   27   47   67   87
#>  [7,]    8   28   48   68   88
#>  [8,]    9   29   49   69   89
#>  [9,]   10   30   50   70   90
#> [10,]   11   31   51   71   91
#> [11,]   12   32   52   72   92
#> [12,]   13   33   53   73   93
#> [13,]   14   34   54   74   94
#> [14,]   15   35   55   75   95
#> [15,]   16   36   56   76   96
#> [16,]   17   37   57   77   97
#> [17,]   18   38   58   78   98
#> [18,]   19   39   59   79   99
#> [19,]   20   40   60   80  100
#> [20,]   21   41   61   81  101